Included Stefanos initial changes to file structure for readability
